+++ src/resolvers/Auth/Mutation/index.ts
@@ -1,0 +1,1 @@
+export { getToken } from "./getToken"
