+++ src/resolvers/Auth/index.ts
@@ -1,0 +1,4 @@
+export * as authMutationResolvers from "./Mutation"
+
+export { default as authMutationTypeDefs } from "./Mutation/typeDefs"
+export { authSharedTypeDefs } from "./sharedTypeDefs"
