--- src/resolvers/Mutation/createAttempt.ts
@@ -1,35 +1,0 @@
-import { ResolverFn } from ".."
-import { User } from "../../entity/User"
-import { Attempt } from "../../entity/Attempt"
-
-type CreateAttemptArgs = {
-  userId: number
-  send: boolean
-  grade: number
-  date: Date
-}
-
-export const createAttempt: ResolverFn<
-  Attempt,
-  CreateAttemptArgs
-> = async function (parent, { userId, send, grade, date }, context, info) {
-  const { connection } = context
-
-  const user = await connection
-    .getRepository(User)
-    .createQueryBuilder("user")
-    .where("user.id = :id", { id: userId })
-    .getOne()
-
-  if (!user) throw new Error("No user found")
-
-  const attempt = new Attempt()
-  attempt.user = user
-  attempt.grade = grade
-  attempt.send = send
-  attempt.date = date ?? new Date()
-
-  await connection.manager.save(attempt)
-
-  return attempt
-}
