--- src/resolvers/Mutation/index.ts
@@ -1,3 +1,0 @@
-export { createAttempt } from "./createAttempt"
-export { getToken } from "./getToken"
-export { s3GetSignedPutObjectUrl } from "./s3GetSignedPutObjectUrl"
