+++ src/resolvers/S3/index.ts
@@ -1,0 +1,3 @@
+export * as s3QueryResolvers from "./Query"
+
+export { default as s3QueryTypeDefs } from "./Query/typeDefs"
